apps/api/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.44
alembic==1.16.5
psycopg[binary]==3.2.10
pydantic-settings==2.11.0
pgvector==0.3.5
python-multipart==0.0.9
openai==1.76.0
PyJWT==2.10.1
bcrypt==4.3.0
email-validator==2.2.0
